Got basic tasking to work for the agent
Can now task an agent from the server to do a process list. It will perform it and send the results back to the server. Still need to implement the actual task queue populate code. Currently you need to hardcode the task queue into the database for it to work. This will be remedied when the front end client is implemented, which will create tasks that get stored in the database for a specific agent.