Implemented scanning
Got scanning commands to work. The agent will accept them and send back valid results. The results are currently only stored in the database and aren't being rendered to the screen, but we will fix that tomorrow. Also fixed a whole lot of bugs. Added a logging system to the dll where it will produce a MemLib.log file in whatever directory kicks it off. currently only logs scans and errors in scans, but will expand that.