Addresses now populate correctly from scans in the admin client
Got most of the scanning working correctly. Need to fix something with the agent. It appears to be crashing - something about vector index being incorrect, but that should be a quick fix. Still need to make it so addresses can be dropped into the editable section and modified, which involves implementing the write command.